requirements.txt
blinker==1.6.2
Bootstrap-Flask==2.2.0
click==8.1.7
Flask==2.2.5
Flask-CKEditor==0.4.6
Flask-WTF==1.1.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Werkzeug==2.2.3
WTForms==3.0.1
Pillow==10.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
requests==2.31.0
urllib3==2.0.6
